add tile info UI
