+++ packages/client/src/layers/network/api/index.ts
@@ -1,3 +1,4 @@
 export { getBlockAtPosition, getEntityAtPosition, getECSBlock, getTerrainBlock } from "./terrain/getBlockAtPosition";
 export { getTerrain } from "./terrain/utils";
 export { getBiome } from "./terrain/getBiome";
+export { getHeight } from "./terrain/getHeight";
