+++ packages/client/src/layers/react/components/common/index.ts
@@ -4,7 +4,8 @@ export { Slot } from "./Slot";
 export { Border } from "./Border";
 export { AbsoluteBorder } from "./AbsoluteBorder";
 export { Crafting } from "./Crafting";
-export { Container, CloseableContainer } from "./Container";
+export { Container } from "./Container";
+export { CloseableContainer } from "./CloseableContainer";
 export { Button } from "./Button";
 export { LoadingBar } from "./LoadingBar";
 export { IconButton } from "./IconButton";
