+++ packages/client/src/layers/react/components/index.ts
@@ -6,7 +6,7 @@
 // import { registerInventory } from "./Inventory";
 // import { registerSidebar } from "./Sidebar";
 // import { registerPlugins } from "./Plugins";
-import { registerTileHover } from "./TileHover";
+import { registerTileInfo } from "./TileInfo";
 // export * from "./common";
 
 export function registerUIComponents() {
@@ -18,5 +18,5 @@ export function registerUIComponents() {
   // registerSidebar();
   // registerComponentBrowser();
   // registerPlugins();
-  registerTileHover();
+  registerTileInfo();
 }
